Update power-monitor.js
Two major changes in this fork:

1) Whatever number the node receives as input (it is always Watts) is passed through on the output as msg.payload.power (along with the usual other items in the JSON).
2) A new state named "idle" exists which passes through msg.payload.power, msg.payload.name, and msg.payload.event ("idle").

The advantage in these changes is that downstream nodes can get access to the input wattage and updates to things like labels. Before, no information would reach downstream nodes if the device (a Dryer for example) was not used since a reboot or update to node-red. Another advantage of passing through the Watts is that a washing machine current cycle can be detected, ie: "Filling", "Draining", "Washing", "Spin Cycle", etc. Each washing machine has its own profile and I am working on code to detect and automate those downstream nodes.